lib: don't add evalFlake args into the top-level lib namespace
evalFlakeArgs is a private function to mkFlake, so there is no reason
expense going over the top level namespace.

Unfortunately, keeping boundaries around future devos-lib, this also
means we cannot introduce a dependency on pkgs, and vie versa we cannot
introduce a dependency on devos-lib lighthartedly (without a clear api)
on pkgs. While this is genuinly useful, we ight need to find a cleaner
and more organized way to implement this in the future.