lib: merge dev into lib
since lib is structured after upstream lib (lists, attrs, strings), and
it is even conceived as lib.makeExtensible, it does not appear
in line to namespace devos specific library additions: not only is the
name `dev` little self explaining, we also have the `os` namespace, for
if library functions are devos specific. The rest are just confenience
functions which might as well ladn upstream one day.